Updates variable processing scripts

+++ common-files/apply-terragrunt-variables.sh
@@ -3,9 +3,15 @@
 if [[ -f "${PWD}/terragrunt.hcl" ]]; then
   terragrunt apply -auto-approve
 else
+  TERRAFORM_DIR=$(find . -name "main.tf" | grep -v ".terraform/modules" | sed -E 's~^./~~g' | sed -E 's~/main.tf$~~g')
   VARIABLES_FILE="${1}"
+
   if [[ -z "${VARIABLES_FILE}" ]]; then
     VARIABLES_FILE="variables.yaml"
+  fi
+
+  if [[ -z "${CREDENTIALS_FILE}" ]]; then
+    CREDENTIALS_FILE="credentials.yaml"
   fi
 
   YQ=$(command -v yq4 || command -v yq)
@@ -19,49 +25,37 @@ else
     exit 1
   fi
 
-  CREDENTIALS_PROPERTIES="../credentials.properties"
-
-  TERRAFORM_DIR=$(find . -name "main.tf" | grep -v ".terraform/modules" | sed -E 's~^./~~g' | sed -E 's~/main.tf$~~g')
+  CREDENTIALS_TFVARS="${TERRAFORM_DIR}/credentials.auto.tfvars"
   TERRAFORM_TFVARS="${TERRAFORM_DIR}/terraform.tfvars"
-
-  if [[ -f "${TERRAFORM_TFVARS}" ]]; then
-    cp "${TERRAFORM_TFVARS}" "${TERRAFORM_TFVARS}.backup"
-    rm "${TERRAFORM_TFVARS}"
-  fi
-
-  if [[ -f "${CREDENTIALS_PROPERTIES}" ]]; then
-    cp "${CREDENTIALS_PROPERTIES}" "${CREDENTIALS_PROPERTIES}.backup"
-    rm "${CREDENTIALS_PROPERTIES}"
-  fi
-  touch "${CREDENTIALS_PROPERTIES}"
-
-  if [[ ! -f "${VARIABLES_FILE}" ]]; then
-    echo "Variables can be provided in a yaml file passed as the first argument"
-    echo ""
-  fi
-
-  TMP_VARIABLES_FILE="${VARIABLES_FILE}.tmp"
-
-  echo "variables: []" > ${TMP_VARIABLES_FILE}
 
   function process_variable () {
     local name="$1"
     local default_value="$2"
     local sensitive="$3"
     local description="$4"
+    local prompt_all="$5"
 
     local variable_name="TF_VAR_${name}"
 
-    environment_variable=$(env | grep "${variable_name}" | sed -E 's/.*=(.*).*/\1/g')
+    if env | grep -q "${variable_name}"; then
+      environment_variable=$(env | grep "${variable_name}" | sed -E 's/.*=(.*).*/\1/g')
+    else
+      environment_variable="null"
+    fi
+
     value="${environment_variable}"
     if [[ -f "${VARIABLES_FILE}" ]]; then
-      value=$(cat "${VARIABLES_FILE}" | NAME="${name}" ${YQ} e -o json '.variables[] | select(.name == env(NAME)) | .value // ""' - | jq -c -r '.')
-      if [[ -z "${value}" ]]; then
+      value=$(cat "${VARIABLES_FILE}" | NAME="${name}" ${YQ} e -o json '.variables[] | select(.name == env(NAME)) | .value' - | jq -c -r '.')
+      if [[ "${value}" == "null" ]]; then
         value="${environment_variable}"
       fi
     fi
 
-    while [[ -z "${value}" ]]; do
+    if [[ "${value}" == "null" ]] && [[ "${prompt_all}" != "true" ]]; then
+      value="${default_value}"
+    fi
+
+    while [[ "${value}" == "null" ]]; do
       echo "Provide a value for '${name}':"
       if [[ -n "${description}" ]]; then
         echo "  ${description}"
@@ -71,7 +65,7 @@ else
         sensitive_flag="-s"
       fi
       default_prompt=""
-      if [[ -n "${default_value}" ]]; then
+      if [[ "${default_value}" != "null" ]]; then
         default_prompt="(${default_value}) "
       fi
       read -u 1 ${sensitive_flag} -p "> ${default_prompt}" value
@@ -82,40 +76,68 @@ else
 
     if [[ "${sensitive}" != "true" ]]; then
       echo "${name} = \"${output_value}\"" >> "${TERRAFORM_TFVARS}"
-      NAME="${name}" VALUE="${value}" ${YQ} e -i -P '.variables += [{"name": env(NAME), "value": env(VALUE)}]' "${TMP_VARIABLES_FILE}"
+      if [[ -z "${value}" ]]; then
+        NAME="${name}" VALUE="${value}" ${YQ} e -i -P '.variables += [{"name": env(NAME), "value": ""}]' "${TMP_VARIABLES_FILE}"
+      else
+        NAME="${name}" VALUE="${value}" ${YQ} e -i -P '.variables += [{"name": env(NAME), "value": env(VALUE)}]' "${TMP_VARIABLES_FILE}"
+      fi
     else
-      echo "export ${name}=\"${output_value}\"" >> "${CREDENTIALS_PROPERTIES}"
+      echo "${name} = \"${output_value}\"" >> "${CREDENTIALS_TFVARS}"
+      if [[ -z "${value}" ]]; then
+        NAME="${name}" VALUE="${value}" ${YQ} e -i -P '.variables += [{"name": env(NAME), "value": ""}]' "${TMP_CREDENTIALS_FILE}"
+      else
+        NAME="${name}" VALUE="${value}" ${YQ} e -i -P '.variables += [{"name": env(NAME), "value": env(VALUE)}]' "${TMP_CREDENTIALS_FILE}"
+      fi
     fi
   }
 
-  cat "bom.yaml" | ${YQ} e '.spec.variables[] | .name' - | while read name; do
-    variable=$(cat "bom.yaml" | NAME="${name}" ${YQ} e '.spec.variables[] | select(.name == env(NAME))' -)
+  if [[ ! -f "${TERRAFORM_TFVARS}" ]] && [[ ! -f "${CREDENTIALS_TFVARS}" ]]; then
 
-    default_value=$(echo "${variable}" | ${YQ} e -o json '.defaultValue // ""' - | jq -c -r '.')
-    sensitive=$(echo "${variable}" | ${YQ} e '.sensitive // false' -)
-    description=$(echo "${variable}" | ${YQ} e '.description // ""' -)
+    touch "${TERRAFORM_TFVARS}"
+    touch "${CREDENTIALS_TFVARS}"
 
-    process_variable "${name}" "${default_value}" "${sensitive}" "${description}"
-  done
-
-  cat "${VARIABLES_FILE}" | ${YQ} e '.variables[]' -o json - | jq -c '.' | while read var; do
-    name=$(echo "${var}" | jq -r '.name')
-
-    value=$(echo "${var}" | jq -r '.value // empty')
-    sensitive=$(echo "${var}" | jq -r '.sensitive')
-
-    bom_var=$(cat bom.yaml | ${YQ} e '.spec.variables[]' -o json - | jq --arg NAME "${name}" -c 'select(.name == $NAME)')
-
-    if [[ -z "${bom_var}" ]]; then
-      process_variable "${name}" "${value}" "${sensitive}" ""
+    if [[ ! -f "${VARIABLES_FILE}" ]]; then
+      echo "Variables can be provided in a yaml file passed as the first argument"
+      echo ""
     fi
-  done
 
-  cp "${TMP_VARIABLES_FILE}" "${VARIABLES_FILE}"
-  rm "${TMP_VARIABLES_FILE}"
+    TMP_VARIABLES_FILE="${VARIABLES_FILE}.tmp"
+    TMP_CREDENTIALS_FILE="${CREDENTIALS_FILE}.tmp"
 
-  # shellcheck source=../credentials.properties
-  source "${CREDENTIALS_PROPERTIES}"
+    echo "variables: []" > ${TMP_VARIABLES_FILE}
+    echo "variables: []" > ${TMP_CREDENTIALS_FILE}
+
+    cat "bom.yaml" | ${YQ} e '.spec.variables[] | .name' - | while read name; do
+      variable=$(cat "bom.yaml" | NAME="${name}" ${YQ} e '.spec.variables[] | select(.name == env(NAME))' -)
+
+      default_value=$(echo "${variable}" | ${YQ} e -o json '.value' - | jq -c -r '.')
+      sensitive=$(echo "${variable}" | ${YQ} e '.sensitive // false' -)
+      description=$(echo "${variable}" | ${YQ} e '.description // ""' -)
+
+      process_variable "${name}" "${default_value}" "${sensitive}" "${description}" "${PROMPT_ALL}"
+    done
+
+    if [[ -f "${VARIABLES_FILE}" ]]; then
+      cat "${VARIABLES_FILE}" | ${YQ} e '.variables[]' -o json - | jq -c '.' | while read var; do
+        name=$(echo "${var}" | jq -r '.name')
+
+        value=$(echo "${var}" | jq -r '.value // empty')
+        sensitive=$(echo "${var}" | jq -r '.sensitive')
+
+        bom_var=$(cat bom.yaml | ${YQ} e '.spec.variables[]' -o json - | jq --arg NAME "${name}" -c 'select(.name == $NAME)')
+
+        if [[ -z "${bom_var}" ]]; then
+          process_variable "${name}" "${value}" "${sensitive}" "" "${PROMPT_ALL}"
+        fi
+      done
+    fi
+
+    cp "${TMP_VARIABLES_FILE}" "${VARIABLES_FILE}"
+    rm "${TMP_VARIABLES_FILE}"
+
+    cp "${TMP_CREDENTIALS_FILE}" "${CREDENTIALS_FILE}"
+    rm "${TMP_CREDENTIALS_FILE}"
+  fi
 
   cd "${TERRAFORM_DIR}" || exit 1
   terraform init
